add exporting to trial balance

+++ app/Http/Controllers/AccountController.php
@@ -49,7 +49,7 @@ class AccountController extends Controller
                 ];
             })->toArray();
 
-            $filename = 'trial_balance_' . now()->format('Ymd_His') . '.' . ($format === 'csv' ? 'csv' : 'xlsx');
+            $filename = 'trial_balance_' . $start . '_' . $end . '.' . ($format === 'csv' ? 'csv' : 'xlsx');
             $path = storage_path('app/public/exports/' . $filename);
 
             if (!is_dir(dirname($path))) {
@@ -61,6 +61,18 @@ class AccountController extends Controller
                 $fastExcel->configureCsv(';');
             }
             $fastExcel->export($path);
+
+            $content = file_get_contents($path);
+            if ($content === false) {
+                return;
+            }
+
+            if (str_starts_with($content, "\xEF\xBB\xBF")) {
+                return;
+            }
+
+            $content = "\xEF\xBB\xBF" . $content;
+            file_put_contents($path, $content);
 
             return response()->download($path, $filename)->deleteFileAfterSend(true);
         }
